ExploreScreen

+++ Android/app/src/main/java/site/sayaz/ofindex/data/model/Book.kt
@@ -4,13 +4,13 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Book(
-    val bookId: Int,
+    val bookId: Long,
     val name: String? = null,
     val cover: String? = null,
     val author: String? = null,
     val description: String? = null,
     val tag: String? = null,
     val isbn: String? = null,
-    val bookClass: Int? = null
+    val bookClass: Long? = null
 )
